refactor(problem2): 검증 로직과 비즈니스 로직을 별도의 클래스로 분리
